+++ env/env.go
@@ -3,8 +3,6 @@ package env
 import (
 	"os"
 	"strconv"
-
-	"github.com/spf13/cobra"
 )
 
 const MaxDebugLevel = 2
@@ -27,16 +25,13 @@ var debugLevel = &debug{}
 // If LETS_DEBUG set to int - then verbosity is 1 or 2
 // If --debug or -d used multiple times - then verbosity is 1 or 2
 // If -dd used - then verbosity is 2.
-func ParseDebugLevel(cmd *cobra.Command) (int, error) {
-	level, err := cmd.Flags().GetCount("debug")
-	if err != nil {
-		return 0, err
-	}
-
+// When determined - set debug level globally
+func SetDebugLevel(level int) int {
 	if level == 0 {
 		envValue := os.Getenv("LETS_DEBUG")
 
-		level, err = strconv.Atoi(envValue)
+		envLevel, err := strconv.Atoi(envValue)
+		level = envLevel
 
 		if err != nil {
 			// probably not integer, try just determine bool value
@@ -52,7 +47,7 @@ func ParseDebugLevel(cmd *cobra.Command) (int, error) {
 	level = min(level, MaxDebugLevel)
 
 	debugLevel.set(level)
-	return level, nil
+	return level
 }
 
 func min(x, y int) int {