+++ set/set.go
@@ -1,7 +1,5 @@
 package set
 
-// TODOL maybe drop
-
 type Set[T comparable] map[T]struct{}
 
 func (s Set[T]) Add(values ...T) {